+++ common/src/main/java/tocraft/walkers/api/model/EntityUpdaters.java
@@ -5,6 +5,7 @@ import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.entity.state.EntityRenderState;
+import net.minecraft.data.worldgen.DimensionTypes;
 import net.minecraft.tags.BlockTags;
 import net.minecraft.tags.FluidTags;
 import net.minecraft.util.Mth;
@@ -14,6 +15,7 @@ import net.minecraft.world.entity.animal.Parrot;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.dimension.DimensionType;
 import tocraft.walkers.api.model.impl.AbstractHorseEntityUpdater;
 import tocraft.walkers.api.model.impl.ShulkerEntityUpdater;
 import tocraft.walkers.api.model.impl.SquidEntityUpdater;
@@ -143,5 +145,29 @@ public class EntityUpdaters {
         });
 
         EntityUpdaters.register(EntityType.CAT, (player, cat) -> cat.setInSittingPose(false));
+
+        EntityUpdaters.register(EntityType.HOGLIN, (player, hoglin) -> {
+            if (player.level().dimensionType().piglinSafe()) {
+                hoglin.setImmuneToZombification(true);
+            } else {
+                hoglin.setImmuneToZombification(false);
+            }
+        });
+
+        EntityUpdaters.register(EntityType.PIGLIN, (player, piglin) -> {
+            if (player.level().dimensionType().piglinSafe()) {
+                piglin.setImmuneToZombification(true);
+            } else {
+                piglin.setImmuneToZombification(false);
+            }
+        });
+
+        EntityUpdaters.register(EntityType.PIGLIN_BRUTE, (player, piglinBrute) -> {
+            if (player.level().dimensionType().piglinSafe()) {
+                piglinBrute.setImmuneToZombification(true);
+            } else {
+                piglinBrute.setImmuneToZombification(false);
+            }
+        });
     }
 }